LPS-156173 Extract into a new class
